feat: add `selector` flags
refactor: move `select()` out of `Config`

feat: implement ability to override `selector` option
