+++ handlr-regex/src/common/handler.rs
@@ -36,6 +36,7 @@ pub trait Handleable {
         mime_apps: &mut MimeApps,
         system_apps: &SystemApps,
         args: Vec<String>,
+        selector: &str,
         enable_selector: bool,
     ) -> Result<()> {
         self.get_entry()?.exec(
@@ -44,6 +45,7 @@ pub trait Handleable {
             system_apps,
             ExecMode::Open,
             args,
+            selector,
             enable_selector,
         )
     }
@@ -100,6 +102,7 @@ impl DesktopHandler {
         mime_apps: &mut MimeApps,
         system_apps: &SystemApps,
         args: Vec<String>,
+        selector: &str,
         enable_selector: bool,
     ) -> Result<()> {
         self.get_entry()?.exec(
@@ -108,6 +111,7 @@ impl DesktopHandler {
             system_apps,
             ExecMode::Launch,
             args,
+            selector,
             enable_selector,
         )
     }
